LPD-40425 Moved commerce order info item util to commerce-api module
