+++ modules/apps/commerce/commerce-api/src/main/java/com/liferay/commerce/util/CommerceOrderInfoItemUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.order.content.web.internal.info.item.util;
+package com.liferay.commerce.util;
 
 import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.service.CommerceOrderService;
