+++ modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/fragment/renderer/OrderReturnsDataSetFragmentRenderer.java
@@ -8,8 +8,8 @@ package com.liferay.commerce.order.content.web.internal.fragment.renderer;
 import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.model.CommerceReturn;
 import com.liferay.commerce.order.content.web.internal.constants.CommerceOrderFragmentFDSNames;
-import com.liferay.commerce.order.content.web.internal.info.item.util.CommerceOrderInfoItemUtil;
 import com.liferay.commerce.service.CommerceOrderService;
+import com.liferay.commerce.util.CommerceOrderInfoItemUtil;
 import com.liferay.fragment.model.FragmentEntryLink;
 import com.liferay.fragment.renderer.FragmentRenderer;
 import com.liferay.fragment.renderer.FragmentRendererContext;
